+++ packages/moonstone/internal/Picker/Picker.js
@@ -104,8 +104,8 @@ const Picker = class extends React.Component {
 		decrementIcon: React.PropTypes.string,
 
 		/**
-		 * When `true`, the [button]{@glossary button} is shown as disabled and does not
-		 * generate tap [events]{@glossary event}.
+		 * When `true`, the Picker is shown as disabled and does not generate `onChange`
+		 * [events]{@glossary event}.
 		 *
 		 * @type {Boolean}
 		 * @public